SimpleFit 0.95
- Atributo del ojo en el password del LoginActivity
- Hacer un activity independiente para el Register
- Hacer la vinculación de metodos de login

+++ app/src/main/java/com/example/appgimnasiotfg/ui/model/Usuario.kt
@@ -5,5 +5,5 @@ import java.io.Serializable
 data class Usuario(
     var nombre: String = "",
     var altura: Int = 0,
-    var peso: Float = 0f
+    var peso: Double = 0.0
 ) : Serializable
